requirements.txt
Flask==3.1.0
Flask-Mail==0.10.0
python-dotenv==1.0.1
flask-wtf==1.2.2
WTForms==3.2.1
email_validator==2.2.0